formatting

+++ test/unit/productLines/productLines.base.t.sol
@@ -9,7 +9,7 @@ contract ProductLine_Base is EVaultTestBase {
         assertEq(coreProductLine.vaultLookup(address(eTST)), true);
         assertEq(coreProductLine.vaultLookup(vm.addr(100)), false);
         assertEq(coreProductLine.getVaultListLength(), 2);
-        assertEq(coreProductLine.getVaultListSlice(0, type(uint).max)[0], address(eTST));
-        assertEq(coreProductLine.getVaultListSlice(0, type(uint).max)[1], address(eTST2));
+        assertEq(coreProductLine.getVaultListSlice(0, type(uint256).max)[0], address(eTST));
+        assertEq(coreProductLine.getVaultListSlice(0, type(uint256).max)[1], address(eTST2));
     }
 }
